requirements-ultra-simple.txt
# --- API base (pydantic v2) ---
fastapi==0.111.0
starlette==0.37.2
pydantic==2.8.2
pydantic-core==2.20.1

# ASGI
uvicorn==0.30.6
h11==0.14.0
anyio==4.4.0
click==8.1.7

# HTTP
httpx==0.27.2
requests==2.32.3

# Supabase SDK 2.x (compatível com httpx 0.27)
supabase==2.3.4

# misc
python-dotenv==1.0.1
typing-extensions==4.12.2
gunicorn==22.0.0
